Total rewrite. It works with new hardware as a web adapter.
This code translates between the arduino serial protocol from
control_helper.ino, and the web adapter protocol from the house-monitor
project.

The JSON for an adapter looks like:

"control": {
            "button": {
                "black": {},
                "block_1": {},
                "block_2": {},
                "block_3": {},
                "block_4": {},
                "desktop_doorbell": {},
                "red": {}
            },
            "rgb": {
                "block_1_backlight": [0, 0, 0],
                "block_2_backlight": [0, 0, 0],
                "block_3_backlight": [0, 0, 0],
                "block_4_backlight": [0, 0, 0]
            },
            "rgb_target": {}
        },

If a button is pushed, the server side button push URL is invoked. The
backlight color values in the web adapter are kept current. If the backlight
targets are updated, a change request will be passed to the arduino.

In general, I don't like how this turned out, but it's working.